add tests

+++ token/token.go
@@ -3,5 +3,5 @@ package token
 type Token interface {
 	GetTokenName() string
 	GetAddressAsString() string
-	GetDecimals() uint
+	GetDecimals() uint8
 }
